requirements.txt
click==8.1.3
Flask==2.1.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
